backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "AJ Wong",email = "ajwong408@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.3, <4.0"
dependencies = [
    "alembic (>=1.16.5,<2.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "debugpy (>=1.8.16,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "openai (>=1.107.3,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jobspy (>=1.1.82,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "watchdog (>=6.0.0,<7.0.0)"
]

[dependency-groups]
dev = [
    "black (>=25.1.0,<26.0.0)",
    "mypy (>=1.18.1,<2.0.0)",
    "ruff (>=0.13.0,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
